\name{CNOlistToy2}
\docType{data}
\alias{CNOlistToy2}
\title{Toy data with 2 time points}
\description{This data object contains the data associated with the Toy Model example from the package vignette, already loaded and formatted as a CNOlist object, and modified to contain 2 time points.  The second time point is such a way that all  of the signals stay as in time 1, except for cJun and Jnk which go to zero.}
\usage{CNOlistToy}
\format{CNOlistToy is a list with the fields "namesCues" (character vector), "namesStimuli" (character vector), "namesInhibitors" (character vector), "namesSignals" (character vector), "timeSignals" (numerical vector), "valueCues" (numerical matrix), "valueInhibitors" (numerical matrix), "valueStimuli"(numerical matrix), "valueSignals"(numerical matrix). }
\source{This data and model is extracted from the Matlab version of CellNOpt1.0 (\url{http://www.ebi.ac.uk/saezrodriguez/software.html#CellNetOptimizer}).}
\references{J. Saez-Rodriguez, L. G. Alexopoulos, J. Epperlein, R. Samaga, D. A.
Lauffenburger, S. Klamt and P. K. Sorger. Discrete logic modeling as a means to
link protein signaling networks with functional analysis of mammalian signal transduction, Molecular Systems Biology, 5:331, 2009.}
\keyword{datasets}
