\name{CNOlistToyMMB}
\docType{data}
\alias{CNOlistToyMMB}
\title{Toy data}
\description{This data object contains the data associated with the Toy Model example from the package vignette, already loaded and formatted as a CNOlist object.}
\usage{CNOlistToyMMB}
\format{CNOlistToyMMB is a list with the fields "namesCues" (character vector), "namesStimuli" (character vector), "namesInhibitors" (character vector), "namesSignals" (character vector), "timeSignals" (numerical vector), "valueCues" (numerical matrix), "valueInhibitors" (numerical matrix), "valueStimuli"(numerical matrix), "valueSignals"(numerical matrix). }
\source{This data and model is extracted from the Matlab version of CellNOpt1.0 (\url{http://www.ebi.ac.uk/saezrodriguez/software.html#CellNetOptimizer}).}
\references{J. Saez-Rodriguez, L. G. Alexopoulos, J. Epperlein, R. Samaga, D. A. Lauffenburger, S. Klamt and P. K. Sorger. Discrete logic modeling as a means to link protein signaling networks with functional analysis of mammalian signal transduction, Molecular Systems Biology, 5:331, 2009.}
\keyword{datasets}
\keyword{internal}
