\name{cutSimList}
\alias{cutSimList}
\title{
Cut a model structure according to a bitstring
}
\description{
This function is for developers only. 
}
\usage{
cutModel(Model, bitString)
}

\arguments{
  \item{Model}{the Model object to cut.}
  \item{bitString}{the bitString to be used to cut the SimList object.}
}


\value{
  \item{cutModel}{the new Model object}
}

\author{
 T. Cokelaer
}

