\name{cutSimList}
\alias{cutSimList}
\title{
Cut a simList structure according to a bitstring
}
\description{
This function is for developers only. 
}
\usage{
cutSimList(SimList, bitString)
}

\arguments{
  \item{SimList}{the SimList object to cut.}
  \item{bitString}{the bitString to be used to cut the SimList object.}
}


\value{
  \item{cutSimList}{the new simList object}
}

\author{
 T. Cokelaer
}

