\name{graph2sif}
\alias{graph2sif}

\title{
todo
}

\description{
  This function converts a network form sif format to graph format.
}

\usage{
graph2sif(graph, writeSif=FALSE, filename="Graph"){

}


\arguments{
  \item{graph}{
todo
}
  \item{writeSif}{
todo
}
  \item{filename}{
todo
}

}


\details{
todo
}

\value{
\item{sifFile}{todo}
}

\author{
    F.Eduati
}

\seealso{
\link{graph2sif}
}


