\name{inernals}
\alias{internals}
\title{
List of CellNOptR internal functions.
}
\description{
This is a list of functions that are part of CelNOptR but are not exposed to the
end-user. It may be of interest for developers. They may have a manual
associated to it. If so, you can get the documentation in a R console as usual
by tying preceding the name of the function with question tag.
}


 \details{
  \tabular{ll}{
  function: \tab manual \cr
  cSimulator: \tab yes\cr
  buildBitString: \tab yes\cr
  }
}

\author{
T.Cokelaer
}

