\name{pknmodel}
\docType{data}
\alias{pknmodel}
\title{PKN Model}
\description{This data object contains a prior knowledge network as a cellnopt model}
\usage{pknmodel}
\format{pknmodel is a list with fields "reacID" (character vector), "namesSpecies" (character vector), "interMat" (numerical matrix), "notMat"(numerical matrix). }
\source{This data and model is extracted from the Matlab version of CellNOpt1.0 (\url{http://www.ebi.ac.uk/saezrodriguez/software.html#CellNetOptimizer}).}
\keyword{datasets}
