% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnnotatedModel.R
\name{plotAnnotatedModel}
\alias{plotAnnotatedModel}
\title{plots network graph with highlighted edges.

plots the graph of the network and colours the edges based on the provided vector.
it works as plot \code{\link{plotModel}}, but takes extra arguments as explained below.}
\usage{
plotAnnotatedModel(model, CNOlist = NULL, bString = NULL,
  indexIntegr = NULL, signals = NULL, stimuli = NULL,
  inhibitors = NULL, NCNO = NULL, compressed = NULL,
  output = "STDOUT", filename = NULL, graphvizParams = list(),
  show = TRUE, remove_dot = TRUE, annotEdgeColors = c(),
  annotEdgeMatching = c("reacID", "reacPar"), annotNodeColors = NULL)
}
\arguments{
\item{model}{model created by \code{\link{readSIF}} or after \code{\link{preprocessing}}}

\item{CNOlist}{is a CNOlist object}

\item{bString}{a sequence made of numbers between 0 and 1 of same length as
the one returned by the Genetic Algorithm (GA). This is a generalisation of 
the bitString returned by the GA function: several bit strings can be averaged
and used.}

\item{indexIntegr}{additional indices to highlight some edge (optional).}

\item{signals}{a list of nodes belonging to the signals class}

\item{stimuli}{a list of nodes belonging to the stimuli class}

\item{inhibitors}{a list of nodes belonging to the inhibitors class}

\item{NCNO}{a list of NCNO nodes.}

\item{compressed}{a list of compressed nodes}

\item{output}{the type of output (PNG, PDF, SVG accepted)}

\item{filename}{the filename (without extension) used to write the dot file}

\item{graphvizParams}{a list of optional arguments dedicated to Rgraphviz to tune the layout, see \code{\link{plotModel}}}

\item{show}{show the plot (default is True)}

\item{remove_dot}{remove the dot file that has been created. Default is TRUE}

\item{annotEdgeColors}{is a NAMED vetor of color values. Use the string vector found in the model$reacID or
node1name_k_node2name convention to match the colors to the network.}

\item{annotEdgeMatching}{one of c("reacID",'reacPar'); depending which convention was used to name the color values}

\item{annotNodeColors}{is a NAMED vetor of color values for the model nodes. use the NAME of the node.}
}
\description{
plots network graph with highlighted edges.

plots the graph of the network and colours the edges based on the provided vector.
it works as plot \code{\link{plotModel}}, but takes extra arguments as explained below.
}
