% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRules.R
\name{readBND}
\alias{readBND}
\title{Read network from BND file}
\usage{
readBND(filename)
}
\arguments{
\item{filename}{BND file.}
}
\value{
CellNOpt network
}
\description{
BND is a file format used by MaBoSS to store the boolean network definition.
The reader works if the logic for the activation of the node is stated withing the parameter `logic` of the node.
An example file can be found in `https://maboss.curie.fr/pub/example.bnd`.
}
\author{
Luis Tobalina
}
