% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRules.R
\name{readBNET}
\alias{readBNET}
\title{Read network from BNET file}
\usage{
readBNET(filename)
}
\arguments{
\item{filename}{BNET file. The file is a tab delimited file with two columns,
`targets` and `factors`. `factors` contains the logic rule and `targets` the node
activated by the logic rule.}
}
\value{
CellNOpt network
}
\description{
Read network from BNET file
}
\author{
Luis Tobalina
}
