\name{ToyModel2}
\docType{data}
\alias{ToyModel2}
\title{Toy model}
\description{This data object contains the Toy model from the package vignette, already loaded and formatted as a Model object, and modified for the 2 time points version (a negative fedback between cJun and Jnk (!cJun=Jnk) is added).}
\usage{ToyModel}
\format{ToyModel is a list with fields "reacID" (character vector), "namesSpecies" (character vector), "interMat" (numerical matrix), "notMat"(numerical matrix). }
\source{This data and model is extracted from the Matlab version of CellNOpt1.0 (\url{http://www.ebi.ac.uk/saezrodriguez/software.html#CellNetOptimizer}).}
\references{J. Saez-Rodriguez, L. G. Alexopoulos, J. Epperlein, R. Samaga, D. A. Lauffenburger, S. Klamt and P. K. Sorger. Discrete logic modeling as a means to link protein signaling networks with functional analysis of
mammalian signal transduction, Molecular Systems Biology, 5:331, 2009.}
\keyword{datasets}
