\name{cutSimList}
\alias{cutSimList}
\title{
Cut a simList structure according to a bitstring
}
\description{
This function is for developers only. 
}
\usage{
cutSimList(simList, bString)
}

\arguments{
  \item{simList}{the simList object to cut.}
  \item{bString}{the bitString to be used to cut the simList object.}
}


\value{
  \item{cutSimList}{the new simList object}
}

\author{
 T. Cokelaer
}

