\name{inernals}
\alias{internals}
\title{
List of CellNOptR internal functions.
}
\description{
This is a list of function that are part of CelNOptR but are not exposed to the
end-user. It may be of interest for developers. They may have a manual
associated to it.
}


 \details{
  \tabular{ll}{
  function: \tab manual \cr
  cSimulator: \tab yes\cr
  }
}

\author{
T.Cokelaer
}

