\name{internals}
\alias{internals}
\title{
List of CellNOptR internal functions.
}
\description{
This is a list of functions that are part of CelNOptR but are not exposed to the
end-user. It may be of interest for developers. They may have a manual
associated to it. If so, you can get the documentation in a R console as usual
by tying preceding the name of the function with question tag.
}


 \details{
  \tabular{lll}{
  Function:         \tab deprecated since           \tab manual \cr
  ------------         \tab ---------------------          \tab ---------- \cr
  cSimulator:       \tab internal usage             \tab yes    \cr
  buildBitString:   \tab internal usage             \tab yes    \cr
  simulateT1:       \tab 1.3.28 (use simulateTN)   \tab yes     \cr
  }
}

\author{
T.Cokelaer
}

