\name{model2sif}
\alias{model2sif}

\title{
todo
}

\description{
  This function converts a network form sif format to graph format.
}

\usage{
model2sif(Model,optimRes=NA,writeSif=FALSE, filename="Model"){

}


\arguments{
  \item{Model}{
todo
}
  \item{optimRes}{
todo
}
  \item{writeSif}{
todo
}
  \item{filename}{
todo
}

}


\details{
todo
}

\value{
\item{sifFile}{todo}
}

\author{
    F.Eduati
}

\seealso{
\link{model2sif}
}


