% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKN_ToyPB.R
\docType{data}
\name{pknmodel}
\alias{pknmodel}
\title{pknmodel}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(PKN_ToyPB)
}
\description{
Small in-silico case study used for demonstration. Use with CNOlist_ToyPB data.
}
\keyword{datasets}
